Implement a ForecastFactory to build Forecast entities from CERC's API
We take a representation of a Forecast from CERC's API and
build a list of our own Forecast entities.

For now our ForecastBuilder only uses the first zone
returned (i.e. only handles the scenario where we've asked
CERC's API for information on a particular zone). We will
likely develop this to handle all zones: that's what's
returned if we don't request a specific zone e.g.

```
curl "https://cerc.example.com/getforecast/all?numdays=3
```

For now we also leave out the rain and wind information as
we aren't sure how/if/where this is needed.