Use tailwind with a build step from tailwindcss
We now run our development server with `bin/dev` which uses
the new `Procfile.dev`.

This uses `yarn` to run the `build:css` script when a change
in a view is detected. (Tailwind includes only the styles
which are used in our views.)

In order to build our layout and styling we have a new
`tailwind_layout` file and a route/controller/view to work
on, at http://localhost:3000/styled_layout