Refactor turbo stream implementation for updating forecasts
This approach, of returning via a turbo stream template, gives us more flexibility to apply changes to multple distinct elements within the page. In our case, we want to re render the pollen, UV and temperature predictions, but we’ll also want to re render the  alert guidance component in the next commit.