Add styles for incorporating alerts into air pollution tabs
The rules that govern which styles should apply are quite complex, so I have ended up with a slightly verbose implementation that hopefully is at least fairly clear.

The order in which the classes are listed is also important - the ‘inactive’ class needs to be below ‘active’ in order to override it.

There are two versions of each of the DAQI alert classes. This is unfortunate but allows us to simplify the Javascript code that toggles the selected classes.