Document plan for Subscriber API
Users' subscriptions for alerts will be stored by
CERC. In this commit we document our current plan
for:

- verifying emails and phone numbers
- authentication via verified email / phone numbers
- creating new subscriber IDs
- creating subscriptions for a given subscriber ID
- retrieving subscriptions for a given subscriber ID

AUTH FLOW

We plan to verify users' email addresses
and phone numbers by using a notification
service(e.g. GOV.UK Notify) to either email or
TXT a short auth code e.g. `ABC123` to the user.
The user will then be expected to enter this auth
code into a form to verify the delivery mechanism
via which they wish to receive air quality
alerts.

SEQUENCE DIAGRAMS

The sequence diagrams were produced using [Sequence Diagram
for Mac][]. It uses a text format almost identical to that
used in [Mermaid's sequence diagrams][].

It was inspired by [WebSequenceDiagrams][] which is also
very useful. The markup from Sequence Diagram for Mac was
minimally tweaked to conform to the Mermaid syntax in order
to render properly in Github's UI.

[Sequence Diagram for Mac]:
https://macsequencediagram.com

[Mermaid's sequence diagrams]:
https://mermaid.js.org/syntax/sequenceDiagram.html

[WebSequenceDiagrams]:
https://www.websequencediagrams.com