Refactor toggling of prediction guidance to Stimulus
We define a `PredictionController` on the scope of the whole
"prediction wrapper" element. The *controller* has single
*action*: `toggleGuidance` which acts on 3 *targets* within the
prediction:

- `showGuidanceButton`: toggle visibility on the "show" button
- `hideGuidanceButton`: toggle visibility on the "hide" button
- `guidance`: toggle visibility on the guidance itself

This allows us to replace some custom event handlers in
`application.js` with a modular and conventional approach,
which should be more maintainable.

This little toggling behaviour is very similar to the
initial example of using Stimulus in chapter 3 of Noel
Rappin's book "Modern Front-End Development for Rails,
Second Edition", which we recommend.