Run CI on all branches, not just main and develop
Currently when you push to a branch no tests will be run until you open a PR. By changing the CI to run on all branches, we can have it running while writing the PR so that it's done by the time the PR is opened.

The cost is that with multiple pushes we'll be using additional minutes on our GitHub Actions usage, but this is of minimal concern.