Refactor feature test steps
Current we are previously using Cucumber-like method names and defining test steps in a separate file.

While Cucumber syntax makes understanding a scenario slightly easier, it makes it much harder to understand the code itself as we must flick between the spec and the steps. Moving the code back into the spec and adding comments gives us the best of both worlds.

The vast majority of the steps are not reusable, resulting in a lot of additional code for no benefit. Those steps that are reuseable, we move to a forecast helper.