Combine ForecastFactory with CercForecastService
Currently we have two classes, the CercForecastService and the ForecastFactory. CercForecastService is responsible for returning the forecasts from the cache, and refreshing the cache when it is stale. It calls the CERC API and then passes the response to the ForecastFactory.

CercForecastService is the only user of ForecastFactory, and is a very small class, so there is no good reason to have separate class for what is essentially core functionality of the service – building a forecast.

This commit merges ForecastFactory into CercForecastService.