Add zones to map
Each zone consist of a boundary line and a label. We import all the zones from the index file and then add them and the labels (markers) to the map.

Level 1 zones are shown with thick outlines and bold labels.
Level 2 zones are shown with thick outlines and normal labels.
Level 2 zones in London are hidden until a higher level of zoom.