Add search functionality to the map
This uses Maptiler's Geocoding control, and all the behaviour and styling is handled by Maptiler.

Search is restricted to the UK. We could restrict further to roughly Airtext's coverage, but that requires more complex bounding boxes. Instead I've used the proximity setting to make it prefer results that are near London.