Re render entire pollution forecast in Turbo Stream
Now that we are re rendering the tabs as well within our Turbo Stream, we can remove the alert_guidance turbo frame, because this sits within the pollution_forecast partial.

I have also created a separate partial, `selected_day_input`, to capture the currently selected day within the form that updates the forecast to a new location. This would be messy without its own turbo frame and template, because it would require significant inline ERB.

I have also renamed forecast_tabs to pollution_forecast, to distinguish it from the UV, pollen and temperature forecast, and capture the fact that the component does not only consist of tabs (it also includes the alert guidance).