Pass URL parameters for zone and selected day into forecast controller methods
Because we want to be able to navigate directly to a forecast for a given day and zone, we need our show and update methods to be set up to expect these parameters, and to set them as instance variables. We also need to specify default cases where one or both parameters are not specified.