Update day tab component
Before we move our tab change handling from Stimulus to Turbo, we need to be able to calculate the CSS classes for each tab within our view component. When the tabs are re-rendering on every update, we can apply the classes from the server in Rails, in code that we have tested, rather than doing it on the fly in Javascript and relying on more expensive integration test coverage.