Split street map into two layers with custom styles
We replace the stock map styles with two custom styles applied as two separate layers. The first shows the streets and water, and goes under the pollution layer. The second shows place names and goes on top of the pollution layer.

This approach makes the map much more readable.

The appearance of these styles can be managed at https://cloud.maptiler.com/maps/