Split dynamic forecast content in two
Currently we have a single turbo frame that contains all of the forecast content. This allows us to update everything in one go. However, it also means replacing the map unnecessarily. That was fine previously, but now we want to use map panning and zooming to trigger changes to the forecast that won't work anymore.

This change splits the frame into two, a top and a bottom (with the map inbetween them). We use a Turbo stream, triggered by the location selector form to update these frames.