Fix Turbo map caching issue
When switching between the map and other views, Turbo caches the map. This leads to ghost versions of the map being retained, and in some cases the use of the forward and back buttons on the browser not working.

This change stops Turbo from caching the page at all, fixing the problem.