Update forecast when map panned or zoomed
When the map is panned or zoomed, get the new centre coordinates and look them up in the zone polygons. This may return more than one zone where there are overlapping zones of different levels. We choose between them using the zoom.

The forecast is then updated to match that zone.