Change exclamation triangle icon extension to SVG
Currently we are storing SVGs as `.html.erb`. This is confusing and means they don't get included in the asset pipeline.

The existing icon uses a little bit of Ruby logic to change its colour. As part of this change, we move this logic to the component, which also makes it more easily testable.