Add SVG renderer
Currently we are rendering SVGs by changing their extension to `.html.erb`. This is confusing and means they don't get included in the asset pipeline.

However, if we change them to be `.svg` files, the normal `render` method cannot help us. Rails doesn't have an easy mechanism for rendering SVGs, so I have added a helper for this.