Add ability to show dummy forecasts
It's useful to be able to make the app show dummy forecasts so that we can see what it looks like in different conditions during development and testing.

This commit adds an environmental variable flag controlled by a url query parameter `dummy`. If the parameter is set to 1 or 2, the corresponding dummy forecast will be used instead of the real forecast. When the parameter is not set there is no impact.

We use FactoryBot to create the dummy forecast is the same way that we would for a test.