Add alert subscription form
Per design. The user can subscribe to alerts by email, SMS, and voicemail for up to 5 zones.

Currently the verification code feature does not work as we do not have the details of the system that will send these codes, so for now the codes are just printed to the console.