Add Puffing Billy proxy to stub browser HTTP calls
Puffing Billy allows us to stub HTTP calls in the browser (webmock only does this on the server side). This means we can stub calls to the Airtext API, Maptiler etc.

Since Maptiler requires an API key and we don't want to include that in the test files, we need a proxy like this to test it properly.