Remove Tailwind
Tailwind adds a layer of complexity to the codebase and is non-standard for web development, which means it makes things harder for the developers who will take over this codebase from us. It is bringing us very little value.

This PR removes Tailwind and replaces it with standard SCSS. The one very useful bit of Tailwind is the margin and padding utility classes, so those are added back in.

There are some formatting changes as a result of this change, and probably some bits of Tailwind I've missed, but they will be cleanup up when we align with the new Figma designs.