Handle situations where first forecast is not for today
Previously we assumed the first tab always showed today, but in cases where the forecast has not been updated this may not be true, but we still want to make today's tab active by default.

We now only use the date to determine which tab should be shown as active, and ignore the day parameter.