Use CSS rather than HTML to display button label
Reverts fc838ed96b959fa90440653a21c02d630eeb884d

Switch back from using HTML content to using CSS to display the button label so that we can conditionally show and hide it and sync it better to the state of the details block.

We solve the accessibility problem with a title attribute instead.