Explicitly set zone in dummy forecast
When not set, it'll pick something random and then not appear on the alerts page. Explicitly setting it solves this problem.

As per usual this requires some faffing about with Psych's allowed classes (urgh).

We also update the factory so that it follows the behaviour in the app of setting the zone of the cached forecast to be the zone of the first forecast object.