Use `medium` rather than `mode` to describe email, phone etc.
The current documentation for the Subscriber API uses both ‘mode’ and ‘medium’ to describe the channel (email, voicemail, or SMS) by which an alert is sent. I have standardised to ‘medium’ because ‘mode’ is already in use on the Forecast API to describe the kind of pollutant data that should be returned.

`mode` is retained for the query itself as that's what the API still expects at the moment