Mark fields as required
The checkboxes on the final page of the subscription form are required, but this is not clear to the user and can be confusing because most of the fields on that page are not required.

This change adds the required attribute to the checkboxes and adds a red asterisk to show this.

A change is required to the tests because the browser now handles the first-line validation (there is still validation on the server side, but it's harder to test now).