Rename Subscriptions controller to SubscriptionsCreation
We will be adding a subscription management feature, so this change keeps it clear which controller is which.

It might be possible to do everything in one controller, but Wizard, the gem we are using for the multi-step form makes this harder, so it seems to make more sense to keep the two controllers separate.